Extract api node
