+++ packages/p2p/source/api-node-repository.ts
@@ -2,22 +2,6 @@ import { injectable } from "@mainsail/container";
 import { Contracts } from "@mainsail/contracts";
 
 @injectable()
-export class ApiNode implements Contracts.P2P.ApiNode {
-	public url!: string;
-
-	statusCode?: number;
-	latency?: number;
-
-	constructor() {}
-
-	public init(url: string): ApiNode {
-		this.url = url;
-
-		return this;
-	}
-}
-
-@injectable()
 export class ApiNodeRepository implements Contracts.P2P.ApiNodeRepository {
 	readonly #apiNodes: Map<string, Contracts.P2P.ApiNode> = new Map<string, Contracts.P2P.ApiNode>();
 	readonly #apiNodesPending: Map<string, Contracts.P2P.ApiNode> = new Map<string, Contracts.P2P.ApiNode>();
